+++ src/main/scala/stores/application/RequestProcessingFailed.scala
@@ -1,0 +1,15 @@
+/*
+ * Copyright © 2022-2023 by Pervasive Cats S.r.l.s.
+ *
+ * All Rights Reserved.
+ */
+
+package io.github.pervasivecats
+package stores.application
+
+import stores.ValidationError
+
+case object RequestProcessingFailed extends ValidationError {
+
+  override val message: String = "The request processing execution has failed"
+}